refactored feature flag test for sas create endpoint
